requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
diff-match-patch==20241021
Django==5.2.3
django-cors-headers==4.7.0
django-import-export==4.3.13
django-picklefield==3.3
django-q2==1.8.0
django-redis==6.0.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.2
kaggle==1.7.4.5
mongoengine==0.29.1
msgpack==1.1.1
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pymongo==4.15.0
pynput==1.8.1
pyOpenSSL==25.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
zope.interface==8.0
